chore(routes/items.py): recover text from mistaken config file
